requirements.txt
openai
requests
yfinance
newsapi-python
matplotlib
python-docx
pandas
numpy
scikit-learn
beautifulsoup4
lxml 